requirements.txt
streamlit>=1.32.0
google-generativeai>=0.4.0
PyPDF2>=3.0.0
plotly>=5.15.0
pandas>=2.0.0
